requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-SocketIO==5.1.1
Flask-CORS==3.0.10
Flask-Migrate==3.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.23
Jinja2==3.0.1
itsdangerous==2.0.1
click==8.0.1
eventlet==0.33.0
dnspython==2.1.0
email-validator==1.1.3
alembic==1.7.1
python-engineio==4.2.1
python-socketio==5.4.0
requests==2.31.0
gunicorn==20.1.0
PyJWT==2.1.0
bcrypt==3.2.0
folium==0.12.1
geopy==2.2.0
pandas==1.3.3
numpy==1.21.2
psycopg2-binary==2.9.9 